chore: gci doesn't bring value to linting
The value to have deterministic import orders in our code doesn't seem
that useful considering that imports will automatically be cleaned up
and sorted thanks to the native tooling. This should be good enough
without another opinionated tool complaining that the order isn't _that_
specific way. It's low-impact, and not worth having 2 toolings fight
over this.