fix: executable name does not match release artifacts (#6)
* fix: executable name does not match release artifacts

This change fixes a missing `buildkite-plugin` suffix in the executable
name. When we perform a release, the executable name is named after
the repo.

When the command hook is executed, it's looking for a release with
that name, and if it fails to find a match, it returns a 404 error in
the pipeline.

This change also makes this consitent with other plugin repos.

* chore(goreleaser): update schema to v2

* ci: ensure plugin components are valid

The test step in the GHA workflow is currently commented out. This
will be made active once we begin to develop the real plugin in the near
future.