revert: add Athens Go proxy configuration for docker
Due to this repo being public-facing, it isn't allowed to access the
private composite action in `cultureamp/go-private-modules`. This
results in an error where the go-checks action fails to run because
it is unable to pull the action.

Since this repository is public and not using private deps, we can
undo this change.