revert: add Athens Go proxy configuration for docker  (#8)
* fix: workflow yaml indentation in go-checks

* revert: add Athens Go proxy configuration for docker

Due to this repo being public-facing, it isn't allowed to access the
private composite action in `cultureamp/go-private-modules`. This
results in an error where the go-checks action fails to run because
it is unable to pull the action.

Since this repository is public and not using private deps, we can
undo this change.

* test: update environment variable names
In an earlier commit, the Buildkite environment variable was renamed
to reflect the repo. After this change, the test suite was not running
in CI, hence this wasn't caught in review time.

This change is a follow up to the earlier commit that allows the tests
to pass.