test: catch required environment variables
This change fixes a false-positive test case. Previously, there was a
check for non-nil error when a required environment variable was
missing. This was always passing because a different error was being
returned when the environment variable was not found.

This change specifies the error message we expect to return. This
elminates the false-positive.