fix: param names are kebab-case
Prior to this change, the parameter name was in camelCase, and this
casing appears to transfer over when Buildkite sets up the environment
variables for the plugin.

This change uses kebab case for the parameter, which is the casing
Buildkite expects. In turn, this transforms in UPPER_SNAKE_CASE when
the environment variables are set up.