chore: task runner isn't specific to migrations
Although the primary use of the task runner is for database migrations,
there isn't anything specific to the code that indicates it is only
capable of doing this.

Hence, we will keep the name of the tasks and container generic, and
leave the application up to the people who consume the plugin.