fix: issues with ecs-task-runner during prototyping (#22)
* fix: param names are kebab-case

Prior to this change, the parameter name was in camelCase, and this
casing appears to transfer over when Buildkite sets up the environment
variables for the plugin.

This change uses kebab case for the parameter, which is the casing
Buildkite expects. In turn, this transforms in UPPER_SNAKE_CASE when
the environment variables are set up.

* fix: missing newline makes config and Cloudwatch lines look contiguous

* fix: dereference the pointer to get the value of the exit code

This change fixes a missing pointer dereference so that the exit code of
the task is correctly compared, and the correct message is returned
based on the comparison.

* fix: missing newlines to keep logging legible

* chore: task runner isn't specific to migrations

Although the primary use of the task runner is for database migrations,
there isn't anything specific to the code that indicates it is only
capable of doing this.

Hence, we will keep the name of the tasks and container generic, and
leave the application up to the people who consume the plugin.